requirements.txt
beautifulsoup4==4.12.2
lxml==5.0.0
chardet>=4.0.0
pandas>=1.3.0
requests==2.31.0
tqdm>=4.62.0
aiohttp>=3.8.0
pydantic>=1.8.2
rich==13.7.0
loguru==0.7.2
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
aioresponses>=0.7.0
freezegun>=1.2.0
python-dotenv==1.0.0
httpx>=0.24.0
